refactor: test패키지 main과 동일하게 구조 변경
