+++ src/test/java/kitchenpos/menu/ui/MenuGroupRestControllerTest.java
@@ -1,14 +1,13 @@
-package kitchenpos.ui.menu;
+package kitchenpos.menu.ui;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
-import kitchenpos.menu.application.MenuService;
-import kitchenpos.menu.dto.MenuProductResponse;
-import kitchenpos.menu.dto.MenuRequest;
-import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.menu.dto.MenuProductRequest;
-import kitchenpos.menu.ui.MenuRestController;
+import kitchenpos.menu.application.MenuGroupService;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.dto.MenuGroupRequest;
+import kitchenpos.menu.dto.MenuGroupResponse;
+import kitchenpos.menu.ui.MenuGroupRestController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,7 +19,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.hamcrest.Matchers.hasSize;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -29,9 +27,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-
-@WebMvcTest(MenuRestController.class)
-class MenuRestControllerTest {
+@WebMvcTest(MenuGroupRestController.class)
+class MenuGroupRestControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
@@ -39,39 +36,37 @@ class MenuRestControllerTest {
     ObjectMapper mapper;
 
     @MockBean
-    private MenuService menuService;
+    private MenuGroupService menuGroupService;
 
-    @DisplayName("메뉴 등록")
+
+    @DisplayName("메뉴그룹 등록")
     @Test
     public void create() throws Exception {
-        MenuRequest menuRequest = new MenuRequest("쭈꾸미", 1000, 1L, Arrays.asList(new MenuProductRequest(1L, 10L)));
-        MenuResponse menuResponse = new MenuResponse(1L, "쭈꾸미", 1000, Arrays.asList(new MenuProductResponse(1L, 10L)));
-        given(menuService.create(any())).willReturn(menuResponse);
+        MenuGroupResponse expectedMenuGroup = new MenuGroupResponse(1L, "세일");
+        MenuGroupRequest request = new MenuGroupRequest("세일");
+        given(menuGroupService.create(any())).willReturn(expectedMenuGroup);
 
-        mockMvc.perform(post("/api/menus")
+        mockMvc.perform(post("/api/menu-groups")
                 .contentType(MediaType.APPLICATION_JSON)
-                .content(makeJsonString(menuRequest)))
+                .content(makeJsonString(request)))
                 .andDo(print())
                 .andExpect(status().isCreated());
     }
 
-    @DisplayName("메뉴 리스트")
+    @DisplayName("메뉴그룹 리스트")
     @Test
     public void list() throws Exception {
-        List<MenuResponse> results = Arrays.asList(
-                new MenuResponse(1L, "쭈꾸미", 1000, Arrays.asList(new MenuProductResponse(1L, 10L))),
-                new MenuResponse(2L, "볶음밥", 2000, Arrays.asList(new MenuProductResponse(2L, 10L)))
-        );
+        MenuGroup menuGroup1 = new MenuGroup("그룹1");
+        MenuGroup menuGroup2 = new MenuGroup("그룹2");
+        List<MenuGroup> menuGroups = Arrays.asList(menuGroup1, menuGroup2) ;
+        given(menuGroupService.list()).willReturn(menuGroups);
 
-        given(menuService.list()).willReturn(results);
-
-        mockMvc.perform(get("/api/menus"))
+        mockMvc.perform(get("/api/menu-groups"))
                 .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$").isArray())
-                .andExpect(jsonPath("$", hasSize(2)))
-                .andExpect(jsonPath("$.[0].name").value("쭈꾸미"))
-                .andExpect(jsonPath("$.[1].name").value("볶음밥"));
+                .andExpect(jsonPath("$.[0].name").value("그룹1"))
+                .andExpect(jsonPath("$.[1].name").value("그룹2"));
     }
 
     private String makeJsonString(Object request) throws JsonProcessingException {
