+++ src/test/java/kitchenpos/product/ui/ProductRestControllerTest.java
@@ -1,13 +1,11 @@
-package kitchenpos.ui.menu;
+package kitchenpos.product.ui;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
-import kitchenpos.menu.application.MenuGroupService;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.dto.MenuGroupRequest;
-import kitchenpos.menu.dto.MenuGroupResponse;
-import kitchenpos.menu.ui.MenuGroupRestController;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.ui.ProductRestController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +14,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -27,8 +26,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(MenuGroupRestController.class)
-class MenuGroupRestControllerTest {
+@WebMvcTest(ProductRestController.class)
+class ProductRestControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
@@ -36,37 +35,34 @@ class MenuGroupRestControllerTest {
     ObjectMapper mapper;
 
     @MockBean
-    private MenuGroupService menuGroupService;
+    private ProductService productService;
 
 
-    @DisplayName("메뉴그룹 등록")
+    @DisplayName("상품 등록")
     @Test
     public void create() throws Exception {
-        MenuGroupResponse expectedMenuGroup = new MenuGroupResponse(1L, "세일");
-        MenuGroupRequest request = new MenuGroupRequest("세일");
-        given(menuGroupService.create(any())).willReturn(expectedMenuGroup);
-
-        mockMvc.perform(post("/api/menu-groups")
+        Product expectedProduct = new Product("치킨", new BigDecimal(10_000));
+        given(productService.create(any())).willReturn(expectedProduct);
+        mockMvc.perform(post("/api/products")
                 .contentType(MediaType.APPLICATION_JSON)
-                .content(makeJsonString(request)))
+                .content(makeJsonString(expectedProduct)))
                 .andDo(print())
                 .andExpect(status().isCreated());
     }
 
-    @DisplayName("메뉴그룹 리스트")
+    @DisplayName("상품 리스트")
     @Test
     public void list() throws Exception {
-        MenuGroup menuGroup1 = new MenuGroup("그룹1");
-        MenuGroup menuGroup2 = new MenuGroup("그룹2");
-        List<MenuGroup> menuGroups = Arrays.asList(menuGroup1, menuGroup2) ;
-        given(menuGroupService.list()).willReturn(menuGroups);
-
-        mockMvc.perform(get("/api/menu-groups"))
+        Product product1 = new Product("치킨", new BigDecimal(10_000));
+        Product product2 = new Product("스파게티", new BigDecimal(10_000));
+        List<Product> products = Arrays.asList(product1, product2) ;
+        given(productService.list()).willReturn(products);
+        mockMvc.perform(get("/api/products"))
                 .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$").isArray())
-                .andExpect(jsonPath("$.[0].name").value("그룹1"))
-                .andExpect(jsonPath("$.[1].name").value("그룹2"));
+                .andExpect(jsonPath("$.[0].name").value("치킨"))
+                .andExpect(jsonPath("$.[1].name").value("스파게티"));
     }
 
     private String makeJsonString(Object request) throws JsonProcessingException {
