+++ src/test/java/kitchenpos/table/ui/TableGroupRestControllerTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.table;
+package kitchenpos.table.ui;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
